Control.Iterate.BaseTypes

Description

Defines what types can be used in the SetAlgebra, and what operations those types must support (Iter, Basic, Embed)

Documentation

class Iter f where

Minimal complete definition

nxt, lub

Methods

nxt :: f a b -> Collect (a, b, f a b)

lub :: Ord k => k -> f k b -> Collect (k, b, f k b)

hasNxt :: f a b -> Maybe (a, b, f a b)

hasLub :: Ord k => k -> f k b -> Maybe (k, b, f k b)

haskey :: Ord key => key -> f key b -> Bool

isnull :: f k v -> Bool

lookup :: Ord key => key -> f key rng -> Maybe rng

element :: Ord k => k -> f k v -> Collect ()

class Basic f where

In order to build typed Exp (which are a typed deep embedding) of Set operations, we need to know what kind of basic types of Maps and Sets can be used this way. Every Basic type has a few operations for creating one from a list, for adding and removing key-value pairs, looking up a value given a key. Instances of this algebra are functional in that every key has exactly one value associated with it.

Minimal complete definition

addkv, removekey, domain, range

Methods

addpair :: Ord k => k -> v -> f k v -> f k v

in addpair the new value always prevails, to make a choice use addkv which has a combining function that allows choice.

addkv :: Ord k => (k, v) -> f k v -> (v -> v -> v) -> f k v

use ( old new -> old) if you want the v in (f k v) to prevail, and use ( old new -> new) if you want the v in (k,v) to prevail

removekey :: Ord k => k -> f k v -> f k v

domain :: Ord k => f k v -> Set k

range :: Ord v => f k v -> Set v

data BaseRep f k v where

BaseRep witnesses Basic types. I.e. those types that are instances of both Basic and Iter. Pattern matching against a constructor of type BaseRep, determines which base type. For example data Tag f k v = Tag (BaseRep f k v) (f k v) case Tag MapR x -> -- here we know x :: Map.Map k v

Constructors

MapR :: Basic Map => BaseRep Map k v

SetR :: Basic Sett => BaseRep Sett k ()

ListR :: Basic List => BaseRep List k v

SingleR :: Basic Single => BaseRep Single k v

BiMapR :: (Basic (BiMap v), Ord v) => BaseRep (BiMap v) k v

ViewR :: (Monoid coin, Ord cred, Ord ptr, Ord coin, Ord pool) => Tag coin cred pool ptr k v -> BaseRep (View coin cred pool ptr) k v

data List k v where

Constructors

UnSafeList :: Ord k => [(k, v)] -> List k v

unList :: List k v -> [(k, v)]

fromPairs :: Ord k => (v -> v -> v) -> [(k, v)] -> List k v

normalize :: Ord k => (v -> v -> v) -> [(k, v)] -> [(k, v)]

data Single k v where

Constructors

Single :: k -> v -> Single k v

Fail :: Single k v

SetSingle :: k -> Single k ()

data Sett k v where

Constructors

Sett :: Set k -> Sett k ()

class Embed concrete base | concrete -> base where

Methods

toBase :: concrete -> base

fromBase :: base -> concrete