Control.Iterate.Collect

Documentation

newtype Cont ans x

Constructors

Cont

Fields

runCont :: (x -> ans) -> ans

newtype Collect tuple

Constructors

Collect

Fields

runCollect :: forall ans. ans -> (tuple -> ans -> ans) -> ans

fixAction :: Collect tuple -> ans -> (tuple -> ans -> ans) -> ans

A (Collect t) is completely agnostic over how ts are beging collected. We can make this abstraction concrete by using fixAction.

mapify :: Ord a => Collect (a, b) -> Map a b

listify :: Collect (a, b) -> [(a, b)]

count :: Collect (a, b) -> Int

one :: t -> Collect t

Here are several ways to add a new t to what is being collected.

The one and none interface are used when we want collections with 0 or 1 elements

none :: Collect t

front :: t -> Collect t -> Collect t

The front and rear interface can add to either end of the sequence (both in constant time)

rear :: Collect t -> t -> Collect t

when :: Bool -> Collect ()

Conditional collecting

takeC :: Int -> Collect t -> [t]

isempty :: Collect t -> Bool

nonempty :: Collect t -> Bool

hasElem :: Collect t -> Maybe t

newtype ColPlus tuple

Constructors

ColPlus

Fields

runColPlus :: forall ans. ans -> (tuple -> ans -> ans) -> (ans -> ans -> ans) -> ans

runPlus :: Monoid a => ColPlus t -> a -> (t -> a -> a) -> a