Control.SetAlgebra

Documentation

data List k v

data BiMap v a b

type Bimap k v = BiMap v k v

data Single k v where

Constructors

Single :: k -> v -> Single k v

Fail :: Single k v

SetSingle :: k -> Single k ()

class Basic f where

In order to build typed Exp (which are a typed deep embedding) of Set operations, we need to know what kind of basic types of Maps and Sets can be used this way. Every Basic type has a few operations for creating one from a list, for adding and removing key-value pairs, looking up a value given a key. Instances of this algebra are functional in that every key has exactly one value associated with it.

Minimal complete definition

addkv, removekey, domain, range

Methods

addpair :: Ord k => k -> v -> f k v -> f k v

in addpair the new value always prevails, to make a choice use addkv which has a combining function that allows choice.

addkv :: Ord k => (k, v) -> f k v -> (v -> v -> v) -> f k v

use ( old new -> old) if you want the v in (f k v) to prevail, and use ( old new -> new) if you want the v in (k,v) to prevail

removekey :: Ord k => k -> f k v -> f k v

domain :: Ord k => f k v -> Set k

range :: Ord v => f k v -> Set v

class Iter f where

Minimal complete definition

nxt, lub

Methods

nxt :: f a b -> Collect (a, b, f a b)

lub :: Ord k => k -> f k b -> Collect (k, b, f k b)

hasNxt :: f a b -> Maybe (a, b, f a b)

hasLub :: Ord k => k -> f k b -> Maybe (k, b, f k b)

haskey :: Ord key => key -> f key b -> Bool

isnull :: f k v -> Bool

lookup :: Ord key => key -> f key rng -> Maybe rng

element :: Ord k => k -> f k v -> Collect ()

class Embed concrete base | concrete -> base where

Methods

toBase :: concrete -> base

fromBase :: base -> concrete

class HasExp s t | s -> t where

Basic types are those that can be embedded into Exp. The HasExp class, encodes how to lift a Basic type into an Exp. The function toExp will build a typed Exp for that Basic type. This will be really usefull in the smart constructors.

Methods

toExp :: s -> Exp t

data BaseRep f k v where

BaseRep witnesses Basic types. I.e. those types that are instances of both Basic and Iter. Pattern matching against a constructor of type BaseRep, determines which base type. For example data Tag f k v = Tag (BaseRep f k v) (f k v) case Tag MapR x -> -- here we know x :: Map.Map k v

Constructors

MapR :: Basic Map => BaseRep Map k v

SetR :: Basic Sett => BaseRep Sett k ()

ListR :: Basic List => BaseRep List k v

SingleR :: Basic Single => BaseRep Single k v

BiMapR :: (Basic (BiMap v), Ord v) => BaseRep (BiMap v) k v

ViewR :: (Monoid coin, Ord cred, Ord ptr, Ord coin, Ord pool) => Tag coin cred pool ptr k v -> BaseRep (View coin cred pool ptr) k v

dom :: (Ord k, HasExp s (f k v)) => s -> Exp (Sett k ())

rng :: (Ord k, Ord v) => HasExp s (f k v) => s -> Exp (Sett v ())

dexclude :: (Ord k, Iter g, HasExp s1 (g k ()), HasExp s2 (f k v)) => s1 -> s2 -> Exp (f k v)

drestrict :: (Ord k, HasExp s1 (Sett k ()), HasExp s2 (f k v)) => s1 -> s2 -> Exp (f k v)

rexclude :: (Ord k, Iter g, Ord v, HasExp s1 (f k v), HasExp s2 (g v ())) => s1 -> s2 -> Exp (f k v)

rrestrict :: (Ord k, Iter g, Ord v, HasExp s1 (f k v), HasExp s2 (g v ())) => s1 -> s2 -> Exp (f k v)

unionleft :: (Show k, Show v, Ord k, HasExp s1 (f k v), HasExp s2 (g k v)) => s1 -> s2 -> Exp (f k v)

unionright :: (Ord k, HasExp s1 (f k v), HasExp s2 (g k v)) => s1 -> s2 -> Exp (f k v)

unionplus :: (Ord k, Monoid n, HasExp s1 (f k n), HasExp s2 (g k n)) => s1 -> s2 -> Exp (f k n)

singleton :: Ord k => k -> v -> Exp (Single k v)

setSingleton :: Ord k => k -> Exp (Single k ())

intersect :: (Ord k, Iter f, Iter g, HasExp s1 (f k v), HasExp s2 (g k u)) => s1 -> s2 -> Exp (Sett k ())

subset :: (Ord k, Iter f, Iter g, HasExp s1 (f k v), HasExp s2 (g k u)) => s1 -> s2 -> Exp Bool

keyeq :: (Ord k, Iter f, Iter g, HasExp s1 (f k v), HasExp s2 (g k u)) => s1 -> s2 -> Exp Bool

(◁) :: (Ord k, HasExp s1 (Sett k ()), HasExp s2 (f k v)) => s1 -> s2 -> Exp (f k v)

(⋪) :: (Ord k, Iter g, HasExp s1 (g k ()), HasExp s2 (f k v)) => s1 -> s2 -> Exp (f k v)

(▷) :: (Ord k, Iter g, Ord v, HasExp s1 (f k v), HasExp s2 (g v ())) => s1 -> s2 -> Exp (f k v)

(⋫) :: (Ord k, Iter g, Ord v, HasExp s1 (f k v), HasExp s2 (g v ())) => s1 -> s2 -> Exp (f k v)

(∈) :: (Show k, Ord k, Iter g, HasExp s (g k ())) => k -> s -> Exp Bool

(∉) :: (Show k, Ord k, Iter g, HasExp s (g k ())) => k -> s -> Exp Bool

(∪) :: (Show k, Show v, Ord k, HasExp s1 (f k v), HasExp s2 (g k v)) => s1 -> s2 -> Exp (f k v)

(⨃) :: (Ord k, HasExp s1 (f k v), HasExp s2 (g k v)) => s1 -> s2 -> Exp (f k v)

(∪+) :: (Ord k, Monoid n, HasExp s1 (f k n), HasExp s2 (g k n)) => s1 -> s2 -> Exp (f k n)

(∩) :: (Ord k, Iter f, Iter g, HasExp s1 (f k v), HasExp s2 (g k u)) => s1 -> s2 -> Exp (Sett k ())

(⊆) :: (Ord k, Iter f, Iter g, HasExp s1 (f k v), HasExp s2 (g k u)) => s1 -> s2 -> Exp Bool

(≍) :: (Ord k, Iter f, Iter g, HasExp s1 (f k v), HasExp s2 (g k u)) => s1 -> s2 -> Exp Bool

(<|) :: (Ord k, HasExp s1 (Sett k ()), HasExp s2 (f k v)) => s1 -> s2 -> Exp (f k v)

(|>) :: (Ord k, Iter g, Ord v, HasExp s1 (f k v), HasExp s2 (g v ())) => s1 -> s2 -> Exp (f k v)

(➖) :: (Ord k, Iter f, Iter g, HasExp s1 (f k v), HasExp s2 (g k u)) => s1 -> s2 -> Exp (f k v)

data Exp t where

The self typed GADT: Exp, that encodes the shape of Set expressions. A deep embedding. Exp is a typed Symbolic representation of queries we may ask. It allows us to introspect a query The strategy is to 1) Define Exp so all queries can be represented. 2) Define smart constructors that "parse" the surface syntax, and build a typed Exp 3) Write an evaluate function: eval:: Exp t -> t 4) "eval" can introspect the code and apply efficient domain and type specific translations 5) Use the (Iter f) class to evaluate some Exp that can benefit from its efficient nature.

Constructors

Base :: (Ord k, Basic f, Iter f) => BaseRep f k v -> f k v -> Exp (f k v)

eval :: Embed s t => Exp t -> s

materialize :: Ord k => BaseRep f k v -> Collect (k, v) -> f k v

A witness (BaseRep) can be used to materialize a (Collect k v) into the type witnessed by the BaseRep. Recall a (Collect k v) has no intrinsic type (it is just an ABSTRACT sequence of tuples), so the witness describes how to turn them into the chosen datatype. Note that materialize is meant to be applied to a collection built by iterating over a Query. This produces the keys in ascending order, with no duplicate keys. So we do not need to specify how to merge duplicate values.

biMapToMap :: BiMap v a b -> Map a b

biMapFromMap :: (Ord k, Ord v) => Map k v -> BiMap v k v

biMapFromList :: (Ord k, Ord v) => (v -> v -> v) -> [(k, v)] -> BiMap v k v

biMapEmpty :: BiMap v k v

fromList :: Ord k => BaseRep f k v -> (v -> v -> v) -> [(k, v)] -> f k v

keysEqual :: Ord k => Map k v1 -> Map k v2 -> Bool

forwards :: BiMap v k v -> Map k v

backwards :: BiMap v k v -> Map v (Set k)