Control.Iterate.BaseTypes

Description

Defines what types can be used in the SetAlgebra, and what operations those types must support (Iter, Basic, Embed)

Documentation

class Iter f where

The Set algebra include types that encode finite sets and maps of some type. They have a finite domain, and for each domain element they pair a single range element (unit for sets). We are interested in those finite maps that can iterate their pairs in ascending domain order. The operations are: nxt and lub . lub can skip over many items in sub-linear time, it can make things really fast. Many finite maps can support a support lub operation in sub-linear time. Some examples: Balanced binary trees, Arrays (using binary search), Tries, etc. There are basic and compound Iter instances. Compound types include components with types that have Iter instances.

Minimal complete definition

nxt, lub

Methods

nxt :: f a b -> Collect (a, b, f a b)

lub :: Ord k => k -> f k b -> Collect (k, b, f k b)

hasNxt :: f a b -> Maybe (a, b, f a b)

The next few methods can all be defined via nxt and lub, but for base types there often exist much more efficent means, so the default definitions should be overwritten for such basic types. For compound types with Guards, these are often the only way to define them.

hasLub :: Map k v -> Collect ()

Iter List

Instance details

Defined in Control.Iterate.BaseTypes

Methods

nxt :: List a b -> Collect (a, b, List a b)

lub :: Ord k => k -> List k b -> Collect (k, b, List k b)

hasNxt :: List a b -> Maybe (a, b, List a b)

hasLub :: Ord k => k -> List k b -> Maybe (k, b, List k b)

haskey :: Ord key => key -> List key b -> Bool

isnull :: List k v -> Bool

lookup :: Ord key => key -> List key rng -> Maybe rng

element :: Ord k => k -> List k v -> Collect ()

Iter Sett

Instance details

Defined in Control.Iterate.BaseTypes

Methods

nxt :: Sett a b -> Collect (a, b, Sett a b)

lub :: Ord k => k -> Sett k b -> Collect (k, b, Sett k b)

hasNxt :: Sett a b -> Maybe (a, b, Sett a b)

hasLub :: Ord k => k -> Sett k b -> Maybe (k, b, Sett k b)

haskey :: Ord key => key -> Sett key b -> Bool

isnull :: Sett k v -> Bool

lookup :: Ord key => key -> Sett key rng -> Maybe rng

element :: Ord k => k -> Sett k v -> Collect ()

Iter Single

Instance details

Defined in Control.Iterate.BaseTypes

Methods

nxt :: Single a b -> Collect (a, b, Single a b)

lub :: Ord k => k -> Single k b -> Collect (k, b, Single k b)

hasNxt :: Single a b -> Maybe (a, b, Single a b)

hasLub :: Ord k => k -> Single k b -> Maybe (k, b, Single k b)

haskey :: Ord key => key -> Single key b -> Bool

isnull :: Single k v -> Bool

lookup :: Ord key => key -> Single key rng -> Maybe rng

element :: Ord k => k -> Single k v -> Collect ()

Iter Query

Instance details

Defined in Control.Iterate.Exp

Methods

nxt :: Query a b -> Collect (a, b, Query a b)

lub :: Ord k => k -> Query k b -> Collect (k, b, Query k b)

hasNxt :: Query a b -> Maybe (a, b, Query a b)

hasLub :: Ord k => k -> Query k b -> Maybe (k, b, Query k b)

haskey :: Ord key => key -> Query key b -> Bool

isnull :: Query k v -> Bool

lookup :: Ord key => key -> Query key rng -> Maybe rng

element :: Ord k => k -> Query k v -> Collect ()

class Basic f where

In order to build typed Exp (which are a typed deep embedding) of map and set operations, we need to know what kind of basic types can be used this way. Every Basic type has a few operations for creating one from a list, for adding and removing key-value pairs, looking up a value given a key. Instances of this algebra are functional in that every key has exactly one value associated with it.

Minimal complete definition

Map k v -> Set v

Basic List

Instance details

Defined in Control.Iterate.BaseTypes

Methods

addpair :: Ord k => k -> v -> List k v -> List k v

addkv :: Ord k => (k, v) -> List k v -> (v -> v -> v) -> List k v

removekey :: Ord k => k -> List k v -> List k v

domain :: Ord k => List k v -> Set k

range :: Ord v => List k v -> Set v

Basic Sett

Instance details

Defined in Control.Iterate.BaseTypes

Methods

addpair :: Ord k => k -> v -> Sett k v -> Sett k v

addkv :: Ord k => (k, v) -> Sett k v -> (v -> v -> v) -> Sett k v

removekey :: Ord k => k -> Sett k v -> Sett k v

domain :: Ord k => Sett k v -> Set k

range :: Ord v => Sett k v -> Set v

Basic Single

Instance details

Defined in Control.Iterate.BaseTypes

Methods

addpair :: Ord k => k -> v -> Single k v -> Single k v

addkv :: Ord k => (k, v) -> Single k v -> (v -> v -> v) -> Single k v

removekey :: Ord k => k -> Single k v -> Single k v

domain :: Ord k => Single k v -> Set k

range :: Ord v => Single k v -> Set v

embedding

data BaseRep f k v where

BaseRep witnesses Basic types. I.e. those types that are instances of both Basic and Iter. Pattern matching against a constructor of type BaseRep, determines which base type. For example data Tag f k v = Tag (BaseRep f k v) (f k v) case Tag MapR x -> -- here we know x :: Map.Map k v

Constructors

MapR :: Map k v

SetR :: Basic Sett => BaseRep Sett k ()

ListR :: Basic List => BaseRep List k v

SingleR :: Basic Single => BaseRep Single k v

data List k v where

Maps stored as lists. Sorted [(key,value)] pairs, with no duplicate keys. The constructor for List is hidden, since it requires some invariants. Use fromPairs to build an initial List.

Constructors

UnSafeList :: Ord k => [(k, v)] -> List k v

unList :: List k v -> [(k, v)]

fromPairs :: Ord k => (v -> v -> v) -> [(k, v)] -> List k v

normalize :: Ord k => (v -> v -> v) -> [(k, v)] -> [(k, v)]

data Single k v where

Maps and sets with zero or a single pair. Iteration is trivial. Succeeds at most once.

Constructors

Single :: k -> v -> Single k v

Fail :: Single k v

SetSingle :: k -> Single k ()

data Sett k v where

Constructors

Sett :: Set k -> Sett k ()

class Embed concrete base | concrete -> base where

Every iterable type type forms an isomorphism with some Base type. For most Base types the isomorphism is the identity in both directions, but for some, like List and Sett, the embeddings are not the trivial identities because the concrete types are not binary type constructors. The Embed class also allows us to add newtypes which encode some Base type to the system.

Methods

toBase :: concrete -> base Map k v

Embed (Single k v) (Single k v)

Instance details

Defined in Control.Iterate.BaseTypes

Methods

toBase :: Single k v -> Single k v

fromBase :: Single k v -> Single k v